studio-main/backend/requirements/base.txt
Django==4.2.11
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
drf-spectacular==0.27.1
psycopg2-binary==2.9.9
redis==5.0.3
celery==5.3.6
django-celery-beat==2.6.0
django-celery-results==2.5.1
channels==4.0.0
channels-redis==4.2.0
firebase-admin==6.5.0
Pillow==10.2.0
python-decouple==3.8
django-storages==1.14.2
boto3==1.34.34
gunicorn==21.2.0
whitenoise==6.6.0
python-slugify==8.0.4
groq==0.13.0
